Les packets fonctionnent maintenant en données binaires + Continuation du développement du mode multi

+++ src/fr/univ_artois/iut_lens/spaceinvader/network_packet/client/PacketClient.java
@@ -4,7 +4,7 @@ import fr.univ_artois.iut_lens.spaceinvader.network_packet.Packet;
 
 public abstract class PacketClient extends Packet {
 	
-	public PacketClient(int c) {
+	public PacketClient(byte c) {
 		super(c);
 	}
