+++ src/fr/univ_artois/iut_lens/spaceinvader/network_packet/server/PacketServer.java
@@ -4,7 +4,7 @@ import fr.univ_artois.iut_lens.spaceinvader.network_packet.Packet;
 
 public abstract class PacketServer extends Packet {
 
-	public PacketServer(int c) {
+	public PacketServer(byte c) {
 		super(c);
 	}
